Generalised specification of VC sets in run/Makefile
Up till now, the Makefile has always assumed that VC sets are always
characterised using a .lis file provided as argument to a -units
option.  This is slightly cumbersome when working with VCs from a
single unit - one has to create a .lis file with one entry, rather
than simply giving the unit name as a non-option argument to Victor.

The generalisation enables the easy definition of named sets of VCs
from a single unit.  One can select all the VCs from the unit, or just
a single goal or a subrange of goals.

The generalisation continues to support the definition of VC sets for
multiple units, though the syntax in the Makefile has changed
slightly.  

See run/Makefile for documentation and examples.