Corrected unsoundness in inference of variable types in rules
Consider a rule of form

 X/Y >= 1 may_be_deduced_from [X >= 1, Y >= 1].

Previously Victor would observe that X and Y were arguments of an
operator expecting reals, and so would unsoundly give X and Y real
types.  With these fixes, type inference only infers the constraints
that X is int or real, and Y is int or real.

To enable the previous behaviour, use option -assume-var-in-real-pos-is-real.