Updated inference of variable types in rules.
Further instances of potentially unsound type inference were noticed.
Up to this revision, Victor automatically gave a free variable a real
type when all instances of it were arguments of functions expecting a
real.  This is probably normally a reasonable inference to make, but
potentially the rule involving the variable may be sound only for the
variable being an integer.  By default, Victor now fails to type such
variables, requiring the user to explicitly indicate their type using

   goal(checktype(<var>,<type>)) 

assumptions or explicit universal quantifiers.  As a consequence, by
default, Victor now fails to to type almost any rule involving free
variables that ought to have real type.

Options 

  -warn-about-speculative-overload-resolution
  -no-reals
  -user-rule-closure-effort

are no longer recognised, and new options

  -assume-int-or-real-var-is-real
  -assume-int-or-real-var-is-int

along with the option 

  -assume-var-in-real-pos-is-real

introduced in the previous revision, now guide Victor as to whether
and how to apply heuristics for inferring the types of variables that
could be integer or real. 

Now by default Victor will issue a warning message whenever it applies 
one of these heuristics.   To suppress these warnings, use the option

  -suppress-warnings-of-var-type-assumptions.

For more details on variable type inference, see Section 4.2 of the
User Manual.

Make options are provided for selecting these Victor options.  See
run/Makefile for details.