Added support for operations that treat the boolean type as ordered
Occasionally the Examiner outputs rules characterising the
boolean__pos and boolean__val functions, directly corresponding to the
rules characterising *__pos and *__val functions for each enumeration
type *.  These rules involve the succ and pred functions on booleans,
and the <= relation on booleans.  This fix allows Victor to properly
handle these rules.

Also changed the name of option -bit-type-bool-eq-to-iff to -bool-eq-to-iff.  
While the conversion of equalities over booleans to
iffs is useful when the -bit-type option is used to introduce the
formula-term distinction, it doesn't need to be tied to the
introduction of this distinction.  Rearranged the code so one can
use the -bool-eq-to-iff option without the -bit-type option asserted.